Wrap puppeteer-core transparently to allow this module to be used as a drop-in replacement for puppeteer.
